refactor: extract useMobileInteractions hook
